Manage reload_mutex with Reloader
