+++ include/reloader.h
@@ -1,6 +1,7 @@
 #ifndef NEWSBOAT_RELOADER_H_
 #define NEWSBOAT_RELOADER_H_
 
+#include <mutex>
 #include <vector>
 
 namespace newsboat {
@@ -10,6 +11,7 @@ class controller;
 /// \brief Updates feeds (fetches, parses, puts results into controller).
 class Reloader {
 	controller* ctrl;
+	std::mutex reload_mutex;
 
 public:
 	Reloader(controller* c);
@@ -22,6 +24,12 @@ public:
 	///
 	/// If \a indexes is nullptr, all feeds will be reloader.
 	void start_reload_all_thread(std::vector<int>* indexes = nullptr);
+
+	void unlock_reload_mutex()
+	{
+		reload_mutex.unlock();
+	}
+	bool trylock_reload_mutex();
 };
 
 } // namespace newsboat
