+++ mk/mk.deps
@@ -328,7 +328,7 @@ src/reloader.o: src/reloader.cpp include/reloader.h include/controller.h \
  include/configpaths.h include/cliargsparser.h include/feedcontainer.h \
  include/filtercontainer.h include/fslock.h include/regexmanager.h \
  include/reloader.h include/remote_api.h include/urlreader.h \
- include/reloadthread.h include/controller.h
+ include/downloadthread.h include/controller.h include/reloadthread.h
 src/reloadthread.o: src/reloadthread.cpp include/reloadthread.h \
  include/configcontainer.h include/configparser.h include/controller.h \
  include/cache.h include/rss.h include/matcher.h filter/FilterParser.h \
